# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ApiResponse": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"code",
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CostType": {
				"enum": [
					"low",
					"HIGH",
					"MEDIUM"
				],
				"type": "string"
			},
			"RestaurantFiltersDto": {
				"properties": {
					"veg_only": {
						"type": "boolean"
					},
					"cost": {
						"items": {
							"$ref": "#/components/schemas/CostType"
						},
						"type": "array"
					},
					"cusine_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isOpen": {
						"type": "boolean"
					}
				},
				"required": [
					"veg_only",
					"cost",
					"cusine_types",
					"isOpen"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateRestaurantDto": {
				"properties": {
					"restaurant_name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"veg_only": {
						"type": "boolean"
					},
					"cost": {
						"$ref": "#/components/schemas/CostType"
					},
					"cusine_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isOpen": {
						"type": "boolean"
					}
				},
				"required": [
					"restaurant_name",
					"address",
					"veg_only",
					"cost",
					"cusine_types",
					"isOpen"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRestaurantDto": {
				"properties": {
					"restaurant_name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"veg_only": {
						"type": "boolean"
					},
					"cost": {
						"$ref": "#/components/schemas/CostType"
					},
					"cusine_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isOpen": {
						"type": "boolean"
					}
				},
				"required": [
					"restaurant_name",
					"address",
					"veg_only",
					"cost",
					"cusine_types",
					"isOpen"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "restaurant-rest-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "amogh-chavan"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/restaurant/details": {
			"get": {
				"operationId": "GetDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": []
			}
		},
		"/restaurant/details/{id}": {
			"get": {
				"operationId": "FindOneById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/restaurant/filter": {
			"post": {
				"operationId": "FilterRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RestaurantFiltersDto"
							}
						}
					}
				}
			}
		},
		"/restaurant/create": {
			"post": {
				"operationId": "CreateRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRestaurantDto"
							}
						}
					}
				}
			}
		},
		"/restaurant/delete/{id}": {
			"delete": {
				"operationId": "DeleteRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/restaurant/update/{id}": {
			"patch": {
				"operationId": "UpdateRestaurant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Restaurant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRestaurantDto"
							}
						}
					}
				}
			}
		},
		"/app/healthcheck": {
			"get": {
				"operationId": "HealthCheck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Health Check Api"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}